@@ -18,19 +18,36 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
 
-  # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^0.1.2
+  # Service locator
+  get_it: ^7.2.0
+  # Bloc for state management
+  flutter_bloc: ^7.3.3
+  # Value equality
+  equatable: ^2.0.3
+  # Functional programming thingies
+  dartz: ^0.10.0
+  # Remote API
+  internet_connection_checker: ^0.0.1+3
+  # Http calls
+  http: ^0.13.3
+  # Local cache
+  shared_preferences: ^2.0.7
+  #registro
+  logging: ^1.0.2
+
   flutter_zoom_drawer: ^2.1.1
   flutter_screenutil: ^5.0.0+2
   flutter_svg: ^0.21.0+1
+    
 
 dev_dependencies:
   flutter_test:
