@@ -1,4 +1,3 @@
-
 import 'package:bufi_remake/src/models/menu_items.dart';
 import 'package:bufi_remake/src/pages/home_page.dart';
 import 'package:bufi_remake/src/pages/menu_page.dart';
@@ -6,6 +5,7 @@ import 'package:bufi_remake/src/pages/tarjeta_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_zoom_drawer/flutter_zoom_drawer.dart';
+
 void main() {
   runApp(MyApp());
 }
@@ -18,28 +18,19 @@ class MyApp extends StatelessWidget {
       designSize: const Size(375, 812),
       builder: () => MaterialApp(
         debugShowCheckedModeBanner: false,
-        title: 'Flutter Demo',
+        title: 'Bufi',
         theme: ThemeData(
-          // This is the theme of your application.
-          //
-          // Try running your application with "flutter run". You'll see the
-          // application has a blue toolbar. Then, without quitting the app, try
-          // changing the primarySwatch below to Colors.green and then invoke
-          // "hot reload" (press "r" in the console where you ran "flutter run",
-          // or simply save your changes to "hot reload" in a Flutter IDE).
-          // Notice that the counter didn't reset back to zero; the application
-          // is not restarted.
+         
           primarySwatch: Colors.blue,
         ),
-        home: MyHomePage(title: 'Flutter Demo Home Page'),
+        home: MyHomePage (),
       ),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, @required this.title}) : super(key: key);
-  final String title;
+  MyHomePage({Key? key,  }) : super(key: key); 
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -58,14 +49,18 @@ class _MyHomePageState extends State<MyHomePage> {
       style: DrawerStyle.Style1,
       backgroundColor: Colors.blueAccent,
       menuScreen: Builder(
-          builder: (context) => MenuPage(
-              itemSeleccionado: itemSeleccionado,
-              onSelectItem: (item) {
-                setState(() {
-                  itemSeleccionado = item;
-                  ZoomDrawer.of(context).close();
-                });
-              })),
+        builder: (context) => MenuPage(
+          itemSeleccionado: itemSeleccionado,
+          onSelectItem: (item) {
+            setState(
+              () {
+                itemSeleccionado = item;
+                ZoomDrawer.of(context)!.close();
+              },
+            );
+          },
+        ),
+      ),
       mainScreen: obtenerPage(),
     );
   }
@@ -74,7 +69,7 @@ class _MyHomePageState extends State<MyHomePage> {
     switch (itemSeleccionado) {
       case MenuItems.card:
         return TarjetaPage();
-      case MenuItems.card:
+      case MenuItems.home:
         return HomePage();
       default:
         return HomePage();
