@@ -1,9 +1,8 @@
-
 import 'package:bufi_remake/src/widgets/menu_widget.dart';
 import 'package:flutter/material.dart';
 
 class HomePage extends StatelessWidget {
-  const HomePage({ Key key}) : super(key: key);
+  const HomePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -15,6 +14,12 @@ class HomePage extends StatelessWidget {
         elevation: 0,
         centerTitle: true,
         leading: MenuWidget(),
+      ),
+      body: Center(
+        child: Text(
+          'Home',
+          style: TextStyle(color: Colors.white),
+        ),
       ),
     );
   }
