@@ -4,7 +4,7 @@ import 'package:bufi_remake/src/widgets/menu_widget.dart';
 import 'package:flutter/material.dart';
 
 class TarjetaPage extends StatelessWidget {
-  const TarjetaPage({Key key}) : super(key: key);
+  const TarjetaPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -16,6 +16,11 @@ class TarjetaPage extends StatelessWidget {
         elevation: 0,
         centerTitle: true,
         leading: MenuWidget(),
+      ),body: Center(
+        child: Text(
+          'Tarjeta',
+          style: TextStyle(color: Colors.white),
+        ),
       ),
     );
   }
